Revise evolution module for both 2D and 3D
