+++ Apps/2DChannel/lbm2d_hilemms_solid_body.cpp
@@ -39,7 +39,7 @@
 #include <string>
 #include "boundary.h"
 #include "evolution.h"
-#include "evolution3d.h"
+#include "evolution.h"
 #include "flowfield.h"
 #include "hilemms.h"
 #include "model.h"
