+++ Src/flowfield.cpp
@@ -241,7 +241,7 @@ void PrepareFlowField() {
     SetBoundaryNodeType();
 }
 
-void DispResidualError3D(const int iter, const SizeType checkPeriod) {
+void DispResidualError(const int iter, const SizeType checkPeriod) {
     ops_printf("##########Residual Error at %i time step##########\n", iter);
     for (auto& compo : g_Components()) {
         for (auto& macroVar : compo.second.macroVars) {
